# docs/swagger.yaml
definitions:
...
  vo.GameHistories:
    properties:
      address:
        type: string
      bonus:
        type: integer
      cards:
        items:
          type: integer
        type: array
      cards_bonus:
        items:
          type: integer
        type: array
      chosen_index:
        type: integer
      id:
        type: integer
      nick_name:
        type: string
      play_time:
        type: integer
      session:
        type: string
    type: object
  vo.GameHistory:
    properties:
      address:
        type: string
      bonus:
        type: integer
      cards:
        items:
          type: integer
        type: array
      cards_bonus:
        items:
          type: integer
        type: array
      chosen_index:
        type: integer
      id:
        type: integer
      play_time:
        type: integer
      session:
        type: string
    type: object
  vo.GameSession:
    properties:
      address:
        type: string
      bonus:
        type: integer
      cards:
        items:
          type: integer
        type: array
      cards_bonus:
        items:
          type: integer
        type: array
      chosen_index:
        type: integer
      id:
        type: integer
      play_time:
        type: integer
      session:
        type: string
    type: object
...
  vo.Pagination:
    properties:
      curPage:
        type: integer
      items: {}
      totalNum:
        type: integer
    type: object
  vo.PlayGameReq:
    properties:
      address:
        type: string
      cards:
        items:
          type: integer
        type: array
      chosen:
        type: integer
      nonce:
        type: string
      signed_msg:
        type: string
      timestamp:
        type: string
    type: object
  vo.PlayGameResp:
    properties:
      bonus:
        type: integer
      id:
        type: integer
      session:
        type: string
    type: object
...
  vo.Response:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
...
info:
  contact: {}
paths:
...
  /file/upload:
    post:
      consumes:
      - multipart/form-data
      description: uploadFile
      parameters:
      - description: accessToken
        in: header
        name: Authorization
        required: true
        type: string
      - description: file
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/vo.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: UploadFile
      tags:
      - file
  /game/{session}:
    get:
      consumes:
      - application/json
      description: get played game detail
      parameters:
      - description: accessToken
        in: header
        name: Authorization
        required: true
        type: string
      - description: session str
        in: path
        name: session
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/vo.Response'
            - properties:
                data:
                  $ref: '#/definitions/vo.GameSession'
              type: object
      summary: GetGameDetail
      tags:
      - game
  /game/histories:
    get:
      consumes:
      - application/json
      description: get all user played game history
      parameters:
      - description: filter address ,if empty will query all address
        in: query
        name: address
        type: string
      - description: page start from 0,default 0
        in: query
        name: page
        type: integer
      - description: size default 20
        in: query
        name: size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/vo.Response'
            - properties:
                data:
                  allOf:
                  - $ref: '#/definitions/vo.Pagination'
                  - properties:
                      items:
                        items:
                          $ref: '#/definitions/vo.GameHistories'
                        type: array
                    type: object
              type: object
      summary: GetGameHistory
      tags:
      - game
  /game/id/{id}:
    get:
      consumes:
      - application/json
      description: get played game detail by id
      parameters:
      - description: accessToken
        in: header
        name: Authorization
        required: true
        type: string
      - description: session id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/vo.Response'
            - properties:
                data:
                  $ref: '#/definitions/vo.GameSession'
              type: object
      summary: GetGameDetailById
      tags:
      - game
  /game/my/history:
    get:
      consumes:
      - application/json
      description: get my played game history
      parameters:
      - description: accessToken
        in: header
        name: Authorization
        required: true
        type: string
      - description: page start from 0,default 0
        in: query
        name: page
        type: integer
      - description: size default 20
        in: query
        name: size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/vo.Response'
            - properties:
                data:
                  allOf:
                  - $ref: '#/definitions/vo.Pagination'
                  - properties:
                      items:
                        items:
                          $ref: '#/definitions/vo.GameHistory'
                        type: array
                    type: object
              type: object
      summary: GetMyGameHistory
      tags:
      - game
  /game/play:
    post:
      consumes:
      - application/json
      description: play game
      parameters:
      - description: req
        in: body
        name: .
        schema:
          $ref: '#/definitions/vo.PlayGameReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/vo.Response'
            - properties:
                data:
                  $ref: '#/definitions/vo.PlayGameResp'
              type: object
      summary: PlayGame
      tags:
      - game
  /gas/cal:
    get:
      consumes:
      - application/json
      description: calculate gas to ares
      parameters:
      - description: gas
        in: query
        name: gas
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/vo.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: CalGasFee
      tags:
      - gas
...
swagger: "2.0"